Move FSDP helpers to generic utils
[ghstack-poisoned]
